CTS-1052 Send notification in slack for refreshing twitter access token

+++ utils/twitterbot.py
@@ -4,6 +4,7 @@ import random
 
 from django.conf import settings
 from django.db import IntegrityError
+from slackclient import SlackClient
 
 from twitterbot.models import (
     BlackList,
@@ -22,6 +23,18 @@ ACCESS_TOKEN = settings.ACCESS_TOKEN
 ACCESS_TOKEN_SECRET = settings.ACCESS_TOKEN_SECRET
 
 
+def send_message_to_slack(message):
+    sc = SlackClient(settings.SLACK_API_TOKEN)
+    channel = settings.SLACK_CHANNEL
+    sc.api_call(
+        "chat.postMessage",
+        channel=channel,
+        text=message,
+        username='@twitter-notifier'
+    )
+    logger.info('Sent message to slack: {}'.format(message))
+
+
 def follow_users():
     tw_accounts = TargetTwitterAccount.objects.filter(is_follower=False,
                                                       followers_count__gt=800)
@@ -31,15 +44,8 @@ def follow_users():
     api = tweepy.API(auth, wait_on_rate_limit=True,
                      wait_on_rate_limit_notify=True)
 
-    current_user = api.me()
-    if 4801 < current_user.friends_count < 5000:
-        limit = 5000 - current_user.followers_count
-    elif current_user.friends_count >= 5000:
-        # TODO: delete this elif block when 'goformoonshot' account will have
-        # a sufficient balance of friends and subscribers
-        return
-
     limit = random.randrange(350, 650)
+    logger.info("The limit of followers is set to %s", limit)
     counter = 0
     for user in tw_accounts:
 
@@ -47,7 +53,7 @@ def follow_users():
             tw_user = api.get_user(user.user_id)
         except tweepy.error.TweepError as err:
             if err.api_code == 50:
-                print("User {} not found!".format(user.name))
+                logger.info("User {} not found!".format(user.name))
                 try:
                     BlackList.objects.create(user_id=user.user_id,
                                              reason="User not found!")
@@ -55,12 +61,26 @@ def follow_users():
                 except IntegrityError:
                     continue
                 continue
+            elif err.api_code == 89:
+                text = 'Twitter access token is expired. Please,' \
+                       ' refresh it in heroku settings'
+                logger.info(text)
+                send_message_to_slack(text)
             else:
                 raise err
 
         if tw_user.followers_count > 1000:
             logger.info("Follow %s", user)
-            api.create_friendship(tw_user.id)
+            try:
+                api.create_friendship(tw_user.id)
+            except tweepy.error.TweepError as err:
+                if err.api_code == 161:
+                    text = "Unable to follow more people at this time. " \
+                           "Account must have a sufficient balance of " \
+                           "friends and subscribers"
+                    logger.info(text)
+                    send_message_to_slack(text)
+                    return
             user.is_follower = True
             user.save(update_fields=('is_follower', ))
             counter += 1
@@ -121,15 +141,28 @@ def unfollow_users():
 
     # TODO: add logic for getting list of users for unfollowing process
     # list must contain screen_names or user_ids of Twitter User
-    bad_users = TwitterFollower.objects.values_list('user_id', flat=True)[:limit]
+    bad_users = TwitterFollower.objects.values_list('user_id',
+                                                    flat=True)[:limit]
 
     for bad_user in bad_users:
-        bad_user = api.get_user(bad_user)
-        api.destroy_friendship(bad_user.id)
+        try:
+            bad_user = api.get_user(bad_user)
+            result = api.destroy_friendship(bad_user.id)
+        except tweepy.error.TweepError as err:
+            if err.api_code == 50:
+                logger.info("User {} not found!".format(bad_user.name))
+                continue
+            elif err.api_code == 89:
+                text = 'Twitter access token is expired. Please,' \
+                       ' refresh it in Heroku settings'
+                logger.info(text)
+                send_message_to_slack(text)
+                return
 
         # sync our db state due to unfollowing users
-        try:
-            BlackList.objects.create(user_id=bad_user.id)
-            TwitterFollower.objects.filter(user_id=bad_user.id).delete()
-        except IntegrityError:
-            continue
+        if not result.following:
+            try:
+                BlackList.objects.create(user_id=bad_user.id)
+                TwitterFollower.objects.filter(user_id=bad_user.id).delete()
+            except IntegrityError:
+                continue
